Implemented RHI types. Improved graphics pipeline state. Use of d3d constants. Added empty methods on root signature new architecture. Adapted all shaders and render loop with new agnostic design

+++ Code/Engine/Source/Bruno/Core/Memory.h
@@ -7,12 +7,12 @@ namespace Bruno
     inline uint32_t AlignU32(uint32_t valueToAlign, uint32_t alignment)
     {
         alignment -= 1;
-        return (uint32_t)((valueToAlign + alignment) & ~alignment);
+        return valueToAlign + alignment & ~alignment;
     }
 
     inline uint64_t AlignU64(uint64_t valueToAlign, uint64_t alignment)
     {
         alignment -= 1;
-        return (uint64_t)((valueToAlign + alignment) & ~alignment);
+        return valueToAlign + alignment & ~alignment;
     }
 }
